requirements.txt
# Core
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
opencv-python-headless>=4.9.0
hidapi>=0.14.0
numpy>=1.26.0
scipy>=1.12.0

# AI Swing Coach (Phase 6)
anthropic>=0.40.0

# Packaging
py2app>=0.28.0
